libc/newlib: copy headers into sysroot
newlib's make install does not install the headers in the sysroot. When we perform the "final compiler for build" step during a Canadian build, the configure of libstdc++-v3 cannot find the libc headers and fails with "checking for the value of EOF... configure: error: computing EOF failed". Copying the newlib headers into the sysroot fixes this.